+++ src/main/java/menu/Food.java
@@ -1,10 +1,10 @@
 package menu;
 
-public class Food extends Dish {
+public class Food extends Dish{
     private final int recipeID;
-    public Food(int recipeID){
+    public Food(int recipeID) {
         super(recipeID);
-        this.recipeID = recipeID;
+        this.recipeID=recipeID;
     }
 
     /**
@@ -13,17 +13,13 @@ public class Food extends Dish {
      * @return the production cost
      */
     @Override
-    public int getProductionCost() {
-
+    public double getProductionCost() {
         switch (recipeID) {
             case 0 -> productionCost = 400;
-            case 1 -> productionCost = (int) (getIngredientsTotalPrice() * 0.1);
-            case 2 -> productionCost = (int) (getIngredientsTotalPrice() * 0.2);
+            case 1 -> productionCost = (getIngredientsTotalPrice() * 0.1);
+            case 2 -> productionCost = (getIngredientsTotalPrice() * 0.2);
             default -> System.out.println("Food dish not founded");
         }
         return productionCost;
     }
-
-
-
 }
